Fixing OpenAI tests
